+++ src/interpreter/lib/std.ts
@@ -11,7 +11,7 @@ export const std: Record<string, Value> = {
 	'help': STR('SEE: https://github.com/syuilo/aiscript/blob/master/docs/get-started.md'),
 
 	//#region Core
-	'Core:v': STR('0.13.1'), // TODO: package.jsonを参照
+	'Core:v': STR('0.13.2'), // TODO: package.jsonを参照
 
 	'Core:ai': STR('kawaii'),