+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [Read translated version (en)](./translations/en/CHANGELOG.md)
+- `Date:year`,`Date:month`,`Date:day`,`Date:hour`,`Date:minute`,`Date:second`に時間数値の引数を渡して時刻指定可能に
 - array.sortとString用比較関数Str:lt, Str:gtの追加
 # 0.13.1
 - Json:stringifyに関数を渡すと不正な値が生成されるのを修正